Assignment01/requirements.txt
# Các thư viện Python cần thiết cho dự án tự động hóa
# Required Python libraries for the automation project

# Thư viện cơ bản
# Basic libraries
requests>=2.28.2

# Google API
google-api-python-client>=2.70.0
google-auth>=2.16.1
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.3

# Tạo biểu đồ
# Chart generation
matplotlib>=3.7.0
numpy>=1.24.2

# Tạo báo cáo HTML
# HTML report generation
jinja2>=3.1.2

# AI APIs
openai>=1.3.0
anthropic>=0.5.0

# Tiện ích
# Utilities
python-dotenv>=1.0.0
